Improvements in SPIFFs and FATFS
- HAL SPIFFs config now returns the SPIFFs error code.
- When formatting the result code is now returned.
- Update code accordingly.
- Rework build system so that SPIFFs and FATFS are now in the NF assemblies library.
- HAL SPIFFS platform specific is now part of the SPIFFS sources.

+++ targets/AzureRTOS/ST/ORGPAL_PALTHREE/nanoCLR/main.c
@@ -12,7 +12,7 @@
 #include <nanoHAL_v2.h>
 #include <targetPAL.h>
 
-extern uint8_t hal_spiffs_config();
+extern int32_t hal_spiffs_config();
 
 extern CLR_SETTINGS clrSettings;
